top menu + files
added a lot, changed the name from testing --> index, and added contact, about, and news to the files. Non-changing info goes on home, and important, daily news, goes on news. about gives a summary of the page, and contact will save the information given to an sql database on cf.