playing around with PHP, added formspree
began utilising Formspree, rather than attempt to make my own form one, as i would most likely need to use a mix of php and html, and I haven't been able to get PHP/composer to work. idea.png is what I'm going to commission Juligz to do, as I'd really like a proper logo for this website sometime soon. Additionally, when I get the logo, I'll have to figure out how to set it up so that it's on the left of the top title bar/the logo on the pages?