2nd news post done, added to style.
FINALLY coming back to this now that school has started up again. In addition, I have a reason to actually code + work on it, cause it may actually have proper foot traffic now that the Untitled SMP is starting up...... In terms of livestreams on there, I'll most likely do a "live camera feed" and report news in short livestreams. Though, it is a LOT more likely that my news will be done on this website :p Sadly, juligz's civilisation event won't be going forwards, but frankly, moving forwards from that. New article dropped on this too, check it out for more info.